tests/conftest.py
import pytest
from pathlib import Path
from PIL import Image

ASSETS_DIR = Path(__file__).parent / "assets"
BUNDLES_DIR = ASSETS_DIR / "bundles"
ASSETS_INPUT_DIR = ASSETS_DIR / "assets"
MOD_UPDATE_DIR = ASSETS_DIR / "mod_update"
MOD_UPDATE_OLD_DIR = MOD_UPDATE_DIR / "old"
MOD_UPDATE_NEW_DIR = MOD_UPDATE_DIR / "new"


def compare_images_mse(img1: Image.Image, img2: Image.Image) -> float:
    if img1.size != img2.size:
        raise ValueError(f"图片尺寸不匹配: {img1.size} vs {img2.size}")
    
    img1 = img1.convert("RGBA")
    img2 = img2.convert("RGBA")
    
    pixels1 = img1.load()
    pixels2 = img2.load()
    
    width, height = img1.size
    total_diff = 0.0
    
    for y in range(height):
        for x in range(width):
            p1 = pixels1[x, y]
            p2 = pixels2[x, y]
            for c in range(4):
                total_diff += (p1[c] - p2[c]) ** 2
    
    mse = total_diff / (width * height * 4)
    return mse


def find_first_file(directory: Path, extension: str) -> Path | None:
    if not directory.exists():
        return None
    files = list(directory.glob(f"*{extension}"))
    return files[0] if files else None


def has_file(directory: Path, extension: str) -> bool:
    if not directory.exists():
        return False
    return bool(list(directory.glob(f"*{extension}")))


def has_sample_bundle() -> bool:
    return has_file(BUNDLES_DIR, ".bundle")


def has_sample_image() -> bool:
    return has_file(ASSETS_INPUT_DIR, ".png")


def has_sample_skel() -> bool:
    return has_file(ASSETS_INPUT_DIR, ".skel")


def has_mod_update_samples() -> bool:
    return has_file(MOD_UPDATE_OLD_DIR, ".bundle") and has_file(MOD_UPDATE_NEW_DIR, ".bundle")


@pytest.fixture
def sample_bundle_path() -> Path | None:
    return find_first_file(BUNDLES_DIR, ".bundle")


@pytest.fixture
def sample_image_path() -> Path | None:
    return find_first_file(ASSETS_INPUT_DIR, ".png")


@pytest.fixture
def sample_skel_path() -> Path | None:
    return find_first_file(ASSETS_INPUT_DIR, ".skel")


@pytest.fixture
def sample_atlas_path() -> Path | None:
    return find_first_file(ASSETS_INPUT_DIR, ".atlas")


@pytest.fixture
def old_mod_bundle_path() -> Path | None:
    return find_first_file(MOD_UPDATE_OLD_DIR, ".bundle")


@pytest.fixture
def new_original_bundle_path() -> Path | None:
    return find_first_file(MOD_UPDATE_NEW_DIR, ".bundle")
